requirements.txt
arabic-reshaper==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.0
grpcio-tools==1.67.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.13
ultralytics-thop==2.0.9
urllib3==2.2.3
